# perf: ~65% faster `create_custom_fields`