This PR basically extends the benefits of https://github.com/frappe/frappe/pull/13139/ for all execution of `create_custom_fields`

## Performance Improvement

Time spent for creating India Compliance fields

### Before

```python
In [2]: %time create_custom_fields(CUSTOM_FIELDS)
CPU times: user 15.8 s, sys: 458 ms, total: 16.2 s
Wall time: 26.3 s
```


### After

```python
In [2]: %time create_custom_fields(CUSTOM_FIELDS)
CPU times: user 7.04 s, sys: 218 ms, total: 7.26 s
Wall time: 9.11 s
```

## Changes Made

### `create_custom_fields`
- Introduce new flag called `in_create_custom_fields`. Use this flag to disable certain code in `CustomField.on_update`
- Create new Set called `doctypes_to_update`. This is looped over after creating custom fields to execute the disabled code of `CustomField.on_update` only once for each DocType - this was being executed anyway earlier 😄
- Remove redundant condition when setting `ignore_validate` for Setup Wizard.

### `CustomField.on_update`
- Move `clear_cache` to below where validation is being done.
- Check for `frappe.flags.in_create_custom_fields` instead of `frappe.flags.in_setup_wizard`. Should still meet original design requirements of https://github.com/frappe/frappe/pull/13139/

  (I don't mind keeping `frappe.flags.in_setup_wizard` anyway - if needed)

- Remove check for single doctypes - this already happens in `db.updatedb`