@ankush 2022-08-13
> Review tip: [Hide whitespace changes](https://github.blog/changelog/2021-10-14-hiding-whitespace-is-now-remembered-for-each-pull-request/)

Now that we autoformat everything, this is very useful!